chore/doc: auto-drop support for unsupported versions NodeJS
This is a breaking change now, because implicitly one would expect
a build to run in old versions if it did before.

In the future it won't be, by definition. You should always use a
supported version to build your project and the LTS timespans are
graceful enough.

BREAKING CHANGE: this is a change in policy. Support will be droppe